deleting an image now deletes the image from the disk

+++ api/POST/Images/DeleteImages.js
@@ -1,17 +1,27 @@
 import { pool } from "../../database/queries.js";
+import { unlinkSync } from "fs";
 
 export const DeleteImage = (req, res, next) => {
   if (req.session.username && (req.session.isAdmin || req.session.isEditor)) {
-    if (req && req.body && req.body.name) {
+    if (req && req.body && req.body.name && req.body.filename) {
       pool.query(
         "DELETE FROM public.imagegallery WHERE name=$1",
         [req.body.name],
         (error, results) => {
-            console.log(results);
           if (error) {
             res.send({
               status: false,
               message: "QUERY_ERROR",
+            });
+            return;
+          }
+
+          try {
+            unlinkSync(`photos/${req.body.filename}`);
+          } catch (err) {
+            res.send({
+              status: false,
+              message: "IMAGE_NOT_DELETED_FROM_DISK",
             });
             return;
           }
