+++ api/POST/Movies/DeleteMovie.js
@@ -1,8 +1,9 @@
+import { unlinkSync } from "fs";
 import { pool } from "../../database/queries.js";
 
 export const DeleteMovie = (req, res, next) => {
   if (req.session.username && (req.session.isAdmin || req.session.isEditor)) {
-    if (req && req.body && req.body.id) {
+    if (req && req.body && req.body.id && req.body.filename) {
       pool.query(
         "DELETE FROM public.movies WHERE id=$1",
         [req.body.id],
@@ -11,6 +12,16 @@ export const DeleteMovie = (req, res, next) => {
             res.send({
               status: false,
               message: "QUERY_ERROR",
+            });
+            return;
+          }
+
+          try {
+            unlinkSync(`photos/${req.body.filename}`);
+          } catch (err) {
+            res.send({
+              status: false,
+              message: "IMAGE_NOT_DELETED_FROM_DISK",
             });
             return;
           }
